[FEAT] add optional CUDA serialization kernels
