+++ .gitignore
@@ -1,2 +1,4 @@
 data
 exp
+build
+__pycache__
