--- models/utils/serialization/__init__.py
@@ -1,8 +1,0 @@
-from .default import (
-    encode,
-    decode,
-    z_order_encode,
-    z_order_decode,
-    hilbert_encode,
-    hilbert_decode,
-)
